+++ src/main/java/com/internship/dto/UpdateEmployeeRequest.java
@@ -7,6 +7,7 @@ import jakarta.validation.constraints.Past;
 import jakarta.validation.constraints.Size;
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -36,7 +37,7 @@ public class UpdateEmployeeRequest {
     private Gender gender;
 
     @Min(value = 0, message = "salary must be greater than or equal to 0")
-    private Float grossSalary;
+    private BigDecimal grossSalary;
 
     private Long departmentId;
     private Long teamId;
