+++ src/main/java/com/internship/mapper/BonusMapper.java
@@ -5,11 +5,12 @@ import com.internship.entity.Bonus;
 import com.internship.entity.Employee;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Component
 public class BonusMapper {
-    public Bonus toEntity(LocalDate date, Float amount, Employee employee) {
+    public Bonus toEntity(LocalDate date, BigDecimal amount, Employee employee) {
         return Bonus.builder()
                 .bonusDate(date)
                 .amount(amount)
